+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 > Bug Fixes
 - **fzf:** fix <c-y> mapping not loaded
 
+> Code Refactoring
+- **fzf:** refactor mappings
+
 > Features
 - **fzf:** mappings moved to a special location
 - **fzf-lua:** added description to mappings